requirements.txt
contourpy==1.3.0
cycler==0.12.1
fonttools==4.53.1
greenlet==3.0.3
kiwisolver==1.4.5
matplotlib==3.9.2
mysql==0.0.3
mysql-connector-python==9.0.0
mysqlclient==2.2.4
numpy==2.1.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
seaborn==0.13.2
six==1.16.0
SQLAlchemy==2.0.32
typing_extensions==4.12.2
tzdata==2024.1
